Move Literal and Location classes into separate files.
